refactor: Token 레이어 재정의
