+++ backend/src/main/java/com/sjiwon/anotherart/token/application/TokenReissueService.java
@@ -1,10 +1,11 @@
-package com.sjiwon.anotherart.token.service;
+package com.sjiwon.anotherart.token.application;
 
 import com.sjiwon.anotherart.global.exception.AnotherArtException;
 import com.sjiwon.anotherart.global.security.exception.AuthErrorCode;
 import com.sjiwon.anotherart.member.domain.Member;
 import com.sjiwon.anotherart.member.service.MemberFindService;
-import com.sjiwon.anotherart.token.service.response.TokenResponse;
+import com.sjiwon.anotherart.token.application.response.TokenResponse;
+import com.sjiwon.anotherart.token.domain.service.TokenManager;
 import com.sjiwon.anotherart.token.utils.TokenProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
