+++ backend/src/test/java/com/sjiwon/anotherart/common/ServiceTest.java
@@ -8,7 +8,7 @@ import com.sjiwon.anotherart.common.config.RedisTestContainersExtension;
 import com.sjiwon.anotherart.favorite.domain.FavoriteRepository;
 import com.sjiwon.anotherart.member.domain.MemberRepository;
 import com.sjiwon.anotherart.purchase.domain.PurchaseRepository;
-import com.sjiwon.anotherart.token.domain.TokenRepository;
+import com.sjiwon.anotherart.token.domain.repository.TokenRepository;
 import com.sjiwon.anotherart.token.utils.JwtTokenProvider;
 import jakarta.persistence.EntityManager;
 import org.junit.jupiter.api.extension.ExtendWith;
